Fix broken Pause Menu UI
