+++ Assets/Animation/PauseGUI.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cea6509a7b8fd4248ae97592ce583f5b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
