--- Assets/Animation/PauseMenu 1.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: eead1201494896f43aae1cab8bb5160b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
