+++ Assets/Animation/PauseMenu.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 32cfdde9f7e01c848a2fabff873dcd7d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
