--- Assets/Animation/PauseUI.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f54f795abf344b54eb454a5c8cc0136c
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 0
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
