+++ Assets/Scripts/UI/PauseUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 77d93b649fb231d40bf7202ccfa5a1e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
