--- Assets/Scripts/UI/QuiMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 198991efe63fa8c4484fcdf88187d492
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
